// src/components/AUSGSTEZUQuestionnaire/types.ts
export interface AUSEZUTaxType {
  userIsOGEZUMinter: string
  userIsAUSTaxResEZUMintingPeriod?: string
  userLocAUSEZUMintingPeriod?: string
  userRegAUSGSTEZUMintingPeriod?: string
  userIsAUSTaxResident: string
  userInAUS: string
  userRegisteredAUSGST: string
  walletAddresses: string
  declareTruth: string
}